:sparkles:script to download, open, rename, delete matterports
The script associates a name with the matterport id. Furthermore, it scrapes the name of the matterport from the matterport website title so it does not have to be entered manually. The script can manage your matterports by downloading, deleting, renaming and opening the one you select based on their names.